requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.5
multidict==6.0.4
openai==1.3.6
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
slack-bolt==1.18.1
slack-sdk==3.26.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
yarl==1.9.3
